requirements.txt
PyAutoGUI==0.9.53
PySDL2==0.9.14
